requirements.txt
Django==1.6.2
Markdown==2.3.1
Pillow==2.3.0
South==0.8.4
django-crispy-forms==1.4.0
django-debug-toolbar==1.0.1
django-reversion==1.8.0
gunicorn==18.0
ipdb==0.8
ipython==1.1.0
sqlparse==0.1.11
wsgiref==0.1.2
django-tinymce==1.5.2